medadapt/content_server/requirements.txt
mcp>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-dotenv>=0.19.0
xmltodict>=0.12.0
pytest>=7.0.0
pytest-cov>=3.0.0
coverage>=6.3.0 